fix(ci): python-coverage-shards (8) — split oversized test_executor_coverage_edges_part_001.py to satisfy 1500-line maintainability limit
